build type checker as loadable module, not library
The old version of the CMake file build EnerCTypeChecker as a library linking
against libclang. This led to duplicate versions of LLVM and Clang being loaded
at runtime when `clang` was eventually executed. This wouldn't be so bad (just
sloppy) except for the non-reentrant parts of LLVM, wherein stuff like
command-line options are duplicated when this happens.

Note that, contrary to what you'd expect, the clang executable doesn't appear
to use libclang. So linking to that led to just as much duplicate code (or
more!) as linking the necessary static components (*.a) of LLVM and Clang.

This is at odds with the example plugin shipping with Clang:
http://llvm.org/viewvc/llvm-project/cfe/trunk/examples/PrintFunctionNames/CMakeLists.txt?view=markup

The user-visible symptom here consisted of some lines like:
CommandLine Error: Argument 'track-memory' defined more than once!