add "dedorsement" counterpart to endorsements
"Dedorsement" is a silly name I made up for a cast from precise to
approximate. This is useful in a few isolated situations where assignments of
pointer types are involved.

This case in Blacksholes, in particular, has a cast from pointer to "unsigned
long long" and back to pointers. Ugly!