add pass to LLVM's default passes
This lets us run instrumentation/analysis during the invocation of clang. It
removes the need for an invocation of `opt` anywhere in the pipeline. While I
didn't take too much advantage of it yet, this should make it simpler to
build/link without a bunch of crazy shell pipeline stuff. (We do, however,
still need to link in the runtime library for instrumented runs.)