insert alloca instruction at top of function
Andre found a segfault when the alloca was in the inner loop. This makes
sense -- I hadn't realized that an alloca instruction could actually allocate
more space more than once! We now hoist the alloca for the ACCEPT loop counter
to the top (entry block) of the function. This will have performance benefits
too since the stack frame need not be resized.