sandbox-safe output file sequestering
For the benchmarks that output large files (i.e., image benchmarks), we keep
outputs in the filesystem instead of putting them in the DB. That, however,
broke when we stated using sandboxing to isolate different runs of the same
benchmark. This restores that functionality and moves it out of each eval.py,
into experiments.py itself. The eval.py load() method returns a string
starting with 'file:' to indicate the need for a sequester.