fix perforation of while loops
We could previously "preforate" some while loops but not actually remove any
code. This change removes that noise and just doesn't perforate those loops.

This brings up an interesting issue: we *could* perforate any subchain of a
loop, provided we can correctly bypass the chunk of the code in question. No
need for it to be the whole body. This is what we currently do, basically, for
for loops -- we heuristically try to perforate everything but the increment
block.