ugly hack: global reference to transform pass
This gives the alias analysis pass access to the transformation pass. We
probably should figure out how to go through the normal channels to do this,
but it's poorly documented (maybe impossible?) to make passes of different
types (i.e., ImmutablePass and FunctionPass) depend on one another. I've been
bitten by this in the past when trying to use profile information;
sidestepping the issue with an ugly global-variable hack instead.