use strings as opportunity site ids
We were running into trouble because, from one run to the next, the IDs of
sites were changing. This was because (for example) loops were not always
visited in the same order after we optimized one loop. One way to fix this
would have been to decide which loops to optimize in one pass and do the
optimization afterward, but this (using full site descriptions as IDs) seemed
easier, if hackier.