nullify for BBs: avoid "live value deleted" error
To erase an instruction, we now replace it with an undefined value first. This
lets us roll up the whole basic block in order (the dependent instructions
will also be deleted). It also solves any issue with terminators that depend
on deleted instructions; they now just get "false" as their test.