Target-specific runtime library (#17)
The runtime library (previously enercrt, now acceptrt) was built with CMake
once, up front, with the rest of the toolchain. This was a bad idea because
it's platform-specific. Now it's built by app.mk, on demand for benchmarks
that need it.

The output file is tagged with the ARCH variable, so you get
`acceptrt.zynq.bc`. When ARCH is unset, it defaults to `default`, indicating
non-cross-compiling builds. I haven't tested this, but I think this should
also work for `acceptrt.msp430.bc` and such.